refactor: Earn component from jsx to tsx
